+Updated SignUpPage To Add Body On Another File +Added SignUpScreenBody File
